move mock logic into stubs and everything into utils
