+++ src/plugins/explore/public/components/patterns_table/utils/constants.ts
@@ -4,7 +4,7 @@
  */
 
 import { EuiBasicTableColumn } from '@elastic/eui';
-import { PatternItem } from './patterns_table';
+import { PatternItem } from '../patterns_table';
 
 export const PATTERNS_FIELD = 'patterns_field';
 export const COUNT_FIELD = 'count';
@@ -38,44 +38,5 @@ export const patternsTableColumns: Array<EuiBasicTableColumn<PatternItem>> = [
       return val;
     },
     align: 'right',
-  },
-];
-
-// Mock data for patterns
-export const mockPatternItems: PatternItem[] = [
-  {
-    pattern: 'INFO [main] Starting application',
-    ratio: 0.35,
-    count: 350,
-  },
-  {
-    pattern: 'DEBUG [worker-1] Processing request',
-    ratio: 0.25,
-    count: 250,
-  },
-  {
-    pattern: 'INFO [worker-2] Request completed successfully',
-    ratio: 0.15,
-    count: 150,
-  },
-  {
-    pattern: 'WARN [worker-1] Slow query detected',
-    ratio: 0.1,
-    count: 100,
-  },
-  {
-    pattern: 'ERROR [main] Failed to connect to database',
-    ratio: 0.08,
-    count: 80,
-  },
-  {
-    pattern: 'INFO [scheduler] Running scheduled task',
-    ratio: 0.05,
-    count: 50,
-  },
-  {
-    pattern: 'DEBUG [worker-3] Cache hit',
-    ratio: 0.02,
-    count: 20,
   },
 ];
