+++ src/plugins/explore/public/components/patterns_table/utils/utils.tsx
@@ -1,0 +1,4 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
